Cc 53 setupscript (#60)
* Create setup.sh script

This script sets up the required environment variables.
The script checks for the existence of the suggested CUDA PATH and LD_LIBRARY_PATH paths.
If they exist, the script adds them to the PATH and LD_LIBRARY_PATH environment variables.
If either the PATH or the LD_LIBRARY_PATH environment variable are not yet set, the script will set them.
The script relies on the hardcoded version number and architecture (32 or 64 bit), which can be used to enforce versioning.
If the suggested locations cannot be found or set, the user is notified and asked to check their settings.

* Change comment from PATH to LD_LIBRARY_PATH

* Subsume bootstrap.sh into setup.sh

bootstrap.sh is subsumed into setup.sh. Thus, bootstrap.sh can be removed.

* Remove bootstrap.sh

`bootstrap.sh` has been subsumed in `setup.sh`

* Reflect setup.sh

Reflect that the environment setup now uses setup.sh

* Correct typo where to whether.

* Correct number to string

* Remove stray @ symbol

* Remove version number and architecture

* Add back architecture

Looks like the architecture id is not a symlink so it must be included.